Lab7(ITMO)
I was tasked to organize storage of a collection in a relational database (PostgresQL) and eliminate storage of the collection in a file.
To generate the id field, I will use the database's tools (sequence).
The state of the collection in memory will only be updated upon successful addition of an object to the database.
All data retrieval commands should work with the collection in memory, not in the database.
I will organize the ability to register and authorize users. Users have the ability to specify a password.
Passwords will be hashed using the SHA-1 algorithm when stored.
Unauthorized users will be prohibited from executing commands.
When storing objects, information about the user who created the object will be saved.
Users will have the ability to view all objects in the collection, but they can only modify objects that belong to them.
To identify the user, I will send the login and password with each request.